V2.3.2
-replaced the exception system with a quick GUI for adding blocks to either placement animation or particle emission exceptions
-added sliders into the FBP menu, allowing setting precise values
-added a slider for adjusting the density of weather particles
-added a slider for adjusting the ammount of particles emitted by destroying a block
-added a button to disable the notice sound after changing exceptinos 
-added a button to disable block placement particles

-adjusted random fade values
-changed rendering technique - particles now feature realtime shading
-renamed the 'Roll Particles' option to 'Low Traction'

-fixed crash caused by block placement
-fixed block placing animation glitches; with/out optifine
-fixed the random texture glitch when placing or breaking a block
-fixed fancy weather particles falling through rooftops
-fixed menu buttons not having a round hitbox
-fixed chiseled blocks disappearing after placing (Chisels & Bits)
-fixed the shadow under a sapling and other plants during the placing animation
-partially fixed compatibility with ViveCraft mod, only Fancy Weather won't work when ON

-SO MUCH MORE

+++ src_1.12/main/java/com/TominoCZ/FBP/util/FBPMathUtil.java
@@ -1,8 +1,5 @@
 package com.TominoCZ.FBP.util;
 
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.Vec2f;
-
 public class FBPMathUtil {
 	public static double add(double d, double add) {
 		if (d < 0.0D)
@@ -13,6 +10,6 @@ public class FBPMathUtil {
 
 	public static double round(double d, int decimals) {
 		int i = (int) Math.round(d * Math.pow(10, decimals));
-		return ((double) i) / Math.pow(10, decimals);
+		return (i) / Math.pow(10, decimals);
 	}
 }
